+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ## Features
 
 - The server list is now displayed in an orderly manner
+- The command `server list (sl)` now shows the server information (minecraft version, loader, loader version) and the
+  server size
+- The command `backup list` now shows the backup size
 - **`remove-unused-chunks`** command added to remove unused chunks of the world
   using [Potato Peeler Tool](https://github.com/Bottle-M/PotatoPeeler)
     - `potato-peeler.chunk-inhabited-time` configuration option added to specify the minimum time a chunk must be